id field changes

+++ northwind/src/main/java/kodlamaio/northwind/entities/concretes/Category.java
@@ -24,7 +24,7 @@ public class Category {
 	
 	@Id
 	@Column(name = "category_id")
-	private int categoryId;
+	private byte categoryId;
 	
 	@Column(name = "category_name")
 	private String categoryName;
